chore(WIP): likeXMLJetty
- build jetty.xml using code
